specify type of constant

+++ pkg/util/timeofday/time_of_day.go
@@ -34,7 +34,7 @@ const (
 	// Time2400 is a special value to represent the 24:00 input time
 	Time2400 = TimeOfDay(microsecondsPerDay)
 
-	microsecondsPerSecond = 1e6
+	microsecondsPerSecond = int64(1e6)
 	microsecondsPerMinute = 60 * microsecondsPerSecond
 	microsecondsPerHour   = 60 * microsecondsPerMinute
 	microsecondsPerDay    = 24 * microsecondsPerHour
@@ -126,19 +126,19 @@ func (t TimeOfDay) Hour() int {
 	if t == Time2400 {
 		return 24
 	}
-	return int(int64(t)%microsecondsPerDay) / microsecondsPerHour
+	return int(t) % int(microsecondsPerDay) / int(microsecondsPerHour)
 }
 
 // Minute returns the minute offset within the hour specified by t, in the
 // range [0, 59].
 func (t TimeOfDay) Minute() int {
-	return int(int64(t)%microsecondsPerHour) / microsecondsPerMinute
+	return int(t) % int(microsecondsPerHour) / int(microsecondsPerMinute)
 }
 
 // Second returns the second offset within the minute specified by t, in the
 // range [0, 59].
 func (t TimeOfDay) Second() int {
-	return int(int64(t)%microsecondsPerMinute) / microsecondsPerSecond
+	return int(t) % int(microsecondsPerMinute) / int(microsecondsPerSecond)
 }
 
 // Microsecond returns the microsecond offset within the second specified by t,
